@@ -1,5 +1,18 @@
+# Copyright Amazon.com Inc. or its affiliates. All Rights Reserved.
+#
+# Licensed under the Apache License, Version 2.0 (the "License"). You
+# may not use this file except in compliance with the License. A copy of
+# the License is located at
+#
+#     http://aws.amazon.com/apache2.0/
+#
+# or in the "license" file accompanying this file. This file is
+# distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF
+# ANY KIND, either express or implied. See the License for the specific
+# language governing permissions and limitations under the License.
+
 import json
-from typing import Any, Optional
+from typing import Any
 from unittest import mock
 from unittest.mock import Mock, PropertyMock, patch
 
@@ -9,12 +22,15 @@ import pytest
 from braket.aws import AwsDevice, AwsDeviceType, AwsQuantumTask
 from braket.circuits import Circuit
 from braket.device_schema import DeviceActionType
-from braket.device_schema.openqasm_device_action_properties import OpenQASMDeviceActionProperties
+from braket.device_schema.openqasm_device_action_properties import (
+    OpenQASMDeviceActionProperties,
+)
 from braket.device_schema.simulators import GateModelSimulatorDeviceCapabilities
 from braket.devices import LocalSimulator
+from braket.program_sets import ProgramSet
 from braket.simulator import BraketSimulator
-from braket.task_result import GateModelTaskResult
-from braket.tasks import GateModelQuantumTaskResult
+from braket.task_result import GateModelTaskResult, ProgramSetTaskResult
+from braket.tasks import GateModelQuantumTaskResult, ProgramSetQuantumTaskResult
 from pennylane.measurements import MeasurementTransform
 from pennylane.tape import QuantumScript, QuantumTape
 from pennylane.wires import Wires
@@ -33,7 +49,12 @@ ACTION_PROPERTIES = OpenQASMDeviceActionProperties.parse_raw(
             "version": ["1"],
             "supportedOperations": ["rx", "ry", "h", "cy", "cnot", "unitary"],
             "supportedResultTypes": [
-                {"name": "StateVector", "observables": None, "minShots": 0, "maxShots": 0},
+                {
+                    "name": "StateVector",
+                    "observables": None,
+                    "minShots": 0,
+                    "maxShots": 0,
+                },
                 {
                     "name": "AdjointGradient",
                     "observables": ["x", "y", "z", "h", "i"],
@@ -52,14 +73,20 @@ GATE_MODEL_RESULT = GateModelTaskResult(
         "measurements": SNAPSHOTS,
         "measuredQubits": [0, 1],
         "taskMetadata": {
-            "braketSchemaHeader": {"name": "braket.task_result.task_metadata", "version": "1"},
+            "braketSchemaHeader": {
+                "name": "braket.task_result.task_metadata",
+                "version": "1",
+            },
             "id": "task_arn",
             "shots": SHOTS,
             "deviceId": "default",
         },
         "additionalMetadata": {
             "action": {
-                "braketSchemaHeader": {"name": "braket.ir.openqasm.program", "version": "1"},
+                "braketSchemaHeader": {
+                    "name": "braket.ir.openqasm.program",
+                    "version": "1",
+                },
                 "source": "qubit[2] q; cnot q[0], q[1]; measure q;",
             },
         },
@@ -77,14 +104,20 @@ RESULT = GateModelQuantumTaskResult.from_string(
             "resultTypes": [],
             "measuredQubits": [0, 1],
             "taskMetadata": {
-                "braketSchemaHeader": {"name": "braket.task_result.task_metadata", "version": "1"},
+                "braketSchemaHeader": {
+                    "name": "braket.task_result.task_metadata",
+                    "version": "1",
+                },
                 "id": "task_arn",
                 "shots": 1,
                 "deviceId": "default",
             },
             "additionalMetadata": {
                 "action": {
-                    "braketSchemaHeader": {"name": "braket.ir.openqasm.program", "version": "1"},
+                    "braketSchemaHeader": {
+                        "name": "braket.ir.openqasm.program",
+                        "version": "1",
+                    },
                     "source": "qubit[2] q; cnot q[0], q[1]; measure q;",
                 },
             },
@@ -103,14 +136,20 @@ BATCH_RESULT_1 = GateModelQuantumTaskResult.from_string(
             "resultTypes": [],
             "measuredQubits": [0, 1],
             "taskMetadata": {
-                "braketSchemaHeader": {"name": "braket.task_result.task_metadata", "version": "1"},
+                "braketSchemaHeader": {
+                    "name": "braket.task_result.task_metadata",
+                    "version": "1",
+                },
                 "id": "task_arn",
                 "shots": 1,
                 "deviceId": "default",
             },
             "additionalMetadata": {
                 "action": {
-                    "braketSchemaHeader": {"name": "braket.ir.openqasm.program", "version": "1"},
+                    "braketSchemaHeader": {
+                        "name": "braket.ir.openqasm.program",
+                        "version": "1",
+                    },
                     "source": "qubit[2] q; cnot q[0], q[1]; measure q;",
                 },
             },
@@ -128,16 +167,130 @@ BATCH_RESULT_2 = GateModelQuantumTaskResult.from_string(
             "resultTypes": [],
             "measuredQubits": [0, 1],
             "taskMetadata": {
-                "braketSchemaHeader": {"name": "braket.task_result.task_metadata", "version": "1"},
+                "braketSchemaHeader": {
+                    "name": "braket.task_result.task_metadata",
+                    "version": "1",
+                },
                 "id": "task_arn",
                 "shots": 1,
                 "deviceId": "default",
             },
             "additionalMetadata": {
                 "action": {
-                    "braketSchemaHeader": {"name": "braket.ir.openqasm.program", "version": "1"},
+                    "braketSchemaHeader": {
+                        "name": "braket.ir.openqasm.program",
+                        "version": "1",
+                    },
                     "source": "qubit[2] q; cnot q[0], q[1]; measure q;",
                 },
+            },
+        }
+    )
+)
+PROGRAM_RESULT_1 = {
+    "braketSchemaHeader": {
+        "name": "braket.task_result.program_result",
+        "version": "1",
+    },
+    "executableResults": [
+        {
+            "braketSchemaHeader": {
+                "name": "braket.task_result.program_set_executable_result",
+                "version": "1",
+            },
+            "measurements": [SNAPSHOTS[0]],
+            "measuredQubits": [0, 1],
+            "inputsIndex": 0,
+        }
+    ],
+    "source": {
+        "braketSchemaHeader": {
+            "name": "braket.ir.openqasm.program",
+            "version": "1",
+        },
+        "source": "OPENQASM 3.0;\nbit[2] b;\nqubit[2] q;\nh q[0];\ncnot q[0], q[1];\nb[0] = measure q[0];\nb[1] = measure q[1];",  # noqa
+        "inputs": {"theta": [0.12, 2.1]},
+    },
+    "additionalMetadata": {
+        "simulatorMetadata": {
+            "braketSchemaHeader": {
+                "name": "braket.task_result.simulator_metadata",
+                "version": "1",
+            },
+            "executionDuration": 50,
+        }
+    },
+}
+PROGRAM_RESULT_2 = {
+    "braketSchemaHeader": {
+        "name": "braket.task_result.program_result",
+        "version": "1",
+    },
+    "executableResults": [
+        {
+            "braketSchemaHeader": {
+                "name": "braket.task_result.program_set_executable_result",
+                "version": "1",
+            },
+            "measurements": [SNAPSHOTS[1]],
+            "measuredQubits": [0, 1],
+            "inputsIndex": 0,
+        }
+    ],
+    "source": {
+        "braketSchemaHeader": {
+            "name": "braket.ir.openqasm.program",
+            "version": "1",
+        },
+        "source": "OPENQASM 3.0;\nbit[2] b;\nqubit[2] q;\nh q[0];\ncnot q[0], q[1];\nb[0] = measure q[0];\nb[1] = measure q[1];",  # noqa
+        "inputs": {"theta": [0.12, 2.1]},
+    },
+    "additionalMetadata": {
+        "simulatorMetadata": {
+            "braketSchemaHeader": {
+                "name": "braket.task_result.simulator_metadata",
+                "version": "1",
+            },
+            "executionDuration": 50,
+        }
+    },
+}
+PROGRAM_SET_RESULT = ProgramSetQuantumTaskResult.from_object(
+    ProgramSetTaskResult(
+        **{
+            "braketSchemaHeader": {
+                "name": "braket.task_result.program_set_task_result",
+                "version": "1",
+            },
+            "programResults": [PROGRAM_RESULT_1, PROGRAM_RESULT_2],
+            "taskMetadata": {
+                "braketSchemaHeader": {
+                    "name": "braket.task_result.program_set_task_metadata",
+                    "version": "1",
+                },
+                "id": "arn:aws:braket:us-west-2:667256736152:quantum-task/bfebc86f-e4ed-4d6f-8131-addd1a49d6dc",  # noqa
+                "deviceId": "arn:aws:braket:::device/quantum-simulator/amazon/sv1",
+                "requestedShots": 2,
+                "successfulShots": 2,
+                "programMetadata": [{"executables": [{}]}],
+                "deviceParameters": {
+                    "braketSchemaHeader": {
+                        "name": "braket.device_schema.simulators.gate_model_simulator_device_parameters",
+                        "version": "1",
+                    },
+                    "paradigmParameters": {
+                        "braketSchemaHeader": {
+                            "name": "braket.device_schema.gate_model_parameters",
+                            "version": "1",
+                        },
+                        "qubitCount": 2,
+                        "disableQubitRewiring": False,
+                    },
+                },
+                "createdAt": "2024-10-15T19:06:58.986Z",
+                "endedAt": "2024-10-15T19:07:00.382Z",
+                "status": "COMPLETED",
+                "totalFailedExecutables": 0,
             },
         }
     )
@@ -149,6 +302,8 @@ TASK.state.return_value = "COMPLETED"
 TASK_BATCH = Mock()
 TASK_BATCH.results.return_value = [BATCH_RESULT_1, BATCH_RESULT_2]
 type(TASK_BATCH).tasks = PropertyMock(return_value=[TASK, TASK])
+TASK_PROGRAM_SET = Mock()
+TASK_PROGRAM_SET.result.return_value = PROGRAM_SET_RESULT
 
 
 @pytest.mark.xfail(raises=ValueError)
@@ -181,54 +336,52 @@ circs = [Circuit().h(0).cnot(0, 1).rx(0, 0.432).ry(0, 0.543) for s in range(SHOT
 circs[0].h(1)
 
 
+@patch.object(AwsDevice, "properties", new_callable=mock.PropertyMock)
 @patch.object(AwsDevice, "run")
 @patch.object(AwsDevice, "run_batch")
 @pytest.mark.parametrize(
-    "pl_circ, wires, result_types, expected_pl_result",
+    "pl_circ, wires, expected_pl_result",
     [
         (
             CIRCUIT_1,
             2,
-            [
-                {
-                    "type": {
-                        "observable": "x()",
-                        "targets": [[0]],
-                        "parameters": ["p_0", "p_1"],
-                        "type": "expectation_value",
-                    },
-                    "value": {
-                        "expectation": 1.5,
-                    },
-                },
-            ],
             [[1.5]],
         ),
     ],
 )
 @pytest.mark.parametrize(
-    "parallel, expected_braket_circ, return_val, call_count, max_para, max_conn",
+    "parallel, expected_braket_task_spec, return_val, call_count, max_para, max_conn, supports_program_sets",
     [
-        (False, circs, TASK, SHOTS, None, None),
-        (True, circs, TASK_BATCH, 1, 10, 10),
+        (False, circs, TASK, SHOTS, None, None, False),
+        (True, circs, TASK_BATCH, 1, 10, 10, False),
+        (False, ProgramSet(circs), TASK_PROGRAM_SET, 1, None, None, True),
     ],
 )
 def test_shadow_expval_aws_device(
     mock_run_batch,
     mock_run,
+    mock_properties,
     pl_circ,
     wires,
-    result_types,
     expected_pl_result,
     parallel,
-    expected_braket_circ,
+    expected_braket_task_spec,
     return_val,
     call_count,
     max_para,
     max_conn,
+    supports_program_sets,
 ):
+    mock_action = Mock()
+    mock_action.action = {"braket.ir.openqasm.program": None}
+    mock_properties.return_value = mock_action
     dev = _aws_device(
-        wires=2, foo="bar", parallel=parallel, max_parallel=max_para, max_connections=max_conn
+        wires=2,
+        foo="bar",
+        parallel=parallel,
+        max_parallel=max_para,
+        max_connections=max_conn,
+        supports_program_sets=supports_program_sets,
     )
     mock_runner = mock_run_batch if parallel else mock_run
     mock_runner.return_value = return_val
@@ -239,7 +392,7 @@ def test_shadow_expval_aws_device(
     assert res == expected_pl_result[0]
     if parallel:
         mock_runner.assert_called_with(
-            expected_braket_circ,
+            expected_braket_task_spec,
             s3_destination_folder=("foo", "bar"),
             shots=1,
             **kwargs,
@@ -247,8 +400,18 @@ def test_shadow_expval_aws_device(
             poll_interval_seconds=AwsQuantumTask.DEFAULT_RESULTS_POLL_INTERVAL,
             foo="bar",
         )
+    elif supports_program_sets:
+        mock_runner.assert_called_with(
+            expected_braket_task_spec,
+            s3_destination_folder=("foo", "bar"),
+            shots=2,
+            **kwargs,
+            poll_timeout_seconds=AwsQuantumTask.DEFAULT_RESULTS_POLL_TIMEOUT,
+            poll_interval_seconds=AwsQuantumTask.DEFAULT_RESULTS_POLL_INTERVAL,
+            foo="bar",
+        )
     else:
-        for c in expected_braket_circ:
+        for c in expected_braket_task_spec:
             mock_runner.assert_any_call(
                 c,
                 s3_destination_folder=("foo", "bar"),
@@ -262,25 +425,12 @@ def test_shadow_expval_aws_device(
 
 @patch.object(LocalSimulator, "run")
 @pytest.mark.parametrize(
-    "pl_circ, expected_braket_circ, wires, result_types, expected_pl_result",
+    "pl_circ, expected_braket_circ, wires, expected_pl_result",
     [
         (
             CIRCUIT_1,
             Circuit().h(0).cnot(0, 1).rx(0, 0.432).ry(0, 0.543),
             2,
-            [
-                {
-                    "type": {
-                        "observable": "x()",
-                        "targets": [[0]],
-                        "parameters": ["p_0", "p_1"],
-                        "type": "expectation_value",
-                    },
-                    "value": {
-                        "expectation": 1.5,
-                    },
-                },
-            ],
             [[1.5]],
         ),
     ],
@@ -291,7 +441,6 @@ def test_shadow_expval_local(
     pl_circ,
     expected_braket_circ,
     wires,
-    result_types,
     expected_pl_result,
     backend,
 ):
@@ -325,12 +474,12 @@ def _aws_device(
     shots=SHOTS,
     device_arn="baz",
     action_properties=ACTION_PROPERTIES,
+    supports_program_sets=False,
     **kwargs,
 ):
     properties_mock.action = {DeviceActionType.OPENQASM: action_properties}
-    properties_mock.return_value.action.return_value = {
-        DeviceActionType.OPENQASM: action_properties
-    }
+    if supports_program_sets:
+        properties_mock.action[DeviceActionType.OPENQASM_PROGRAM_SET] = action_properties
     type_mock.return_value = device_type
     dev = BraketAwsQubitDevice(
         wires=wires,
@@ -351,7 +500,12 @@ class DummyLocalQubitDevice(BraketQubitDevice):
 
 class DummyCircuitSimulator(BraketSimulator):
     def run(
-        self, program: ir.openqasm.Program, qubits: int, shots: Optional[int], *args, **kwargs
+        self,
+        program: ir.openqasm.Program,
+        qubits: int,
+        shots: int | None,
+        *args,
+        **kwargs,
     ) -> dict[str, Any]:
         self._shots = shots
         self._qubits = qubits
@@ -395,7 +549,7 @@ class DummyCircuitSimulator(BraketSimulator):
                     "supportedResultTypes": [
                         {
                             "name": "resultType1",
-                            "observables": ["observable1"],
+                            "observables": ["z"],
                             "minShots": 2,
                             "maxShots": 4,
                         }
@@ -428,12 +582,16 @@ def test_run_snapshots_not_implemented():
     dev.execute(circuit)
 
 
+@patch.object(AwsDevice, "properties", new_callable=mock.PropertyMock)
 @patch.object(AwsDevice, "run_batch")
-def test_shadows_parallel_tracker(mock_run_batch):
+def test_shadows_parallel_tracker(mock_run_batch, mock_properties):
     """Asserts tracker updates during parallel shadows computation"""
 
     mock_run_batch.return_value = TASK_BATCH
     type(TASK_BATCH).unsuccessful = PropertyMock(return_value={})
+    mock_action = Mock()
+    mock_action.action = {"braket.ir.openqasm.program": None}
+    mock_properties.return_value = mock_action
     dev = _aws_device(wires=2, foo="bar", parallel=True, shots=SHOTS)
 
     with QuantumTape() as circuit:
@@ -463,7 +621,10 @@ def test_shadows_parallel_tracker(mock_run_batch):
 
 class DummyMeasurementTransform(MeasurementTransform):
     def __init__(
-        self, wires: Optional[Wires] = None, seed: Optional[int] = None, id: Optional[str] = None
+        self,
+        wires: Wires | None = None,
+        seed: int | None = None,
+        id: str | None = None,
     ):
         self.seed = seed
         super().__init__(wires=wires, id=id)
