@@ -22,7 +22,6 @@ from braket.pennylane_plugin import PSWAP, CPhaseShift00, CPhaseShift01, CPhaseS
 
 np.random.seed(42)
 
-
 # =========================================================
 
 # list of all non-parametrized single-qubit gates
@@ -85,7 +84,7 @@ class TestHardwareApply:
 
         @qml.qnode(dev)
         def circuit():
-            qml.QubitStateVector.compute_decomposition(state, wires=[0])
+            qml.StatePrep.compute_decomposition(state, wires=[0])
             return qml.probs(wires=range(1))
 
         assert np.allclose(circuit(), np.abs(state) ** 2, **tol)
@@ -177,7 +176,7 @@ class TestHardwareApply:
     def assert_op_and_inverse(op, dev, state, wires, tol, op_args):
         @qml.qnode(dev)
         def circuit():
-            qml.QubitStateVector.compute_decomposition(state, wires=wires)
+            qml.StatePrep.compute_decomposition(state, wires=wires)
             op(*op_args, wires=wires)
             return qml.probs(wires=wires)
 
@@ -185,7 +184,7 @@ class TestHardwareApply:
 
         @qml.qnode(dev)
         def circuit_inv():
-            qml.QubitStateVector.compute_decomposition(state, wires=wires)
+            qml.StatePrep.compute_decomposition(state, wires=wires)
             qml.adjoint(op(*op_args, wires=wires))
             return qml.probs(wires=wires)
 
@@ -197,7 +196,7 @@ class TestHardwareApply:
     def assert_noise_op(op, dev, state, wires, tol, op_args):
         @qml.qnode(dev)
         def circuit():
-            qml.QubitStateVector.compute_decomposition(state, wires=wires)
+            qml.StatePrep.compute_decomposition(state, wires=wires)
             op(*op_args, wires=wires)
             return qml.probs(wires=wires)
 
