@@ -16,6 +16,7 @@
 import pennylane as qml
 import pytest
 from braket.aws import AwsDevice
+from braket.devices import LocalSimulator
 from pennylane import numpy as np
 
 from braket.pennylane_plugin import BraketAwsQubitDevice, BraketLocalQubitDevice
@@ -28,12 +29,9 @@ def test_batch_execution_of_gradient(device, shots, mocker):
     qubits = 2
     layers = 2
 
-    dev_aws = device(qubits)
+    dev_braket = device(qubits)
 
-    if isinstance(dev_aws, BraketLocalQubitDevice):
-        pytest.skip("Parallelized batch execution is only supported on the remote AWS device")
-
-    dev_aws._parallel = True
+    dev_braket._parallel = True
 
     dev_default = qml.device("default.qubit", wires=qubits)
 
@@ -41,24 +39,29 @@ def test_batch_execution_of_gradient(device, shots, mocker):
         qml.templates.StronglyEntanglingLayers(weights, wires=range(qubits))
         return qml.expval(qml.PauliZ(0))
 
-    qnode_aws = qml.QNode(func, dev_aws, diff_method="parameter-shift")
+    qnode_braket = qml.QNode(func, dev_braket, diff_method="parameter-shift")
     qnode_default = qml.QNode(func, dev_default, diff_method="parameter-shift")
 
     shape = qml.templates.StronglyEntanglingLayers.shape(layers, qubits)
     weights = np.random.random(shape)
 
-    dfunc_aws = qml.grad(qnode_aws)
+    dfunc_braket = qml.grad(qnode_braket)
     dfunc_default = qml.grad(qnode_default)
 
-    spy1 = mocker.spy(BraketAwsQubitDevice, "execute")
-    spy2 = mocker.spy(BraketAwsQubitDevice, "batch_execute")
-    spy3 = mocker.spy(AwsDevice, "run_batch")
+    if isinstance(dev_braket, BraketAwsQubitDevice):
+        spy1 = mocker.spy(BraketAwsQubitDevice, "execute")
+        spy2 = mocker.spy(BraketAwsQubitDevice, "batch_execute")
+        spy3 = mocker.spy(AwsDevice, "run_batch")
+    elif isinstance(dev_braket, BraketLocalQubitDevice):
+        spy1 = mocker.spy(BraketLocalQubitDevice, "execute")
+        spy2 = mocker.spy(BraketLocalQubitDevice, "batch_execute")
+        spy3 = mocker.spy(LocalSimulator, "run_batch")
 
-    res_aws = dfunc_aws(weights)
+    res_braket = dfunc_braket(weights)
     res_default = dfunc_default(weights)
 
     if qml.version() >= "0.20.0":
-        assert np.allclose(res_aws, res_default)
+        assert np.allclose(res_braket, res_default)
         spy1.assert_not_called()
         assert len(spy2.call_args_list) == 2
         assert len(spy3.call_args_list) == 2
@@ -69,7 +72,7 @@ def test_batch_execution_of_gradient(device, shots, mocker):
             len(spy2.call_args_list[1][0][1]) == expected_circuits
         )  # Then called for backward pass
     else:
-        assert np.allclose(res_aws, res_default)
+        assert np.allclose(res_braket, res_default)
         spy1.assert_called_once()  # For a forward pass
         spy2.assert_called_once()
         spy3.assert_called_once()
@@ -81,7 +84,8 @@ def test_batch_execution_of_gradient(device, shots, mocker):
 @pytest.mark.parametrize("shots", [None])
 def test_batch_execution_of_gradient_torch(device, shots, mocker):
     """Test that the output of a parallelized execution of batch circuits to evaluate the
-    gradient is correct in comparison to default.qubit when using the torch interface."""
+    gradient is correct in comparison to default.qubit when using the torch interface.
+    """
     try:
         import torch
     except ImportError:
@@ -90,12 +94,9 @@ def test_batch_execution_of_gradient_torch(device, shots, mocker):
     qubits = 2
     layers = 2
 
-    dev_aws = device(qubits)
+    dev_braket = device(qubits)
 
-    if isinstance(dev_aws, BraketLocalQubitDevice):
-        pytest.skip("Parallelized batch execution is only supported on the remote AWS device")
-
-    dev_aws._parallel = True
+    dev_braket._parallel = True
 
     dev_default = qml.device("default.qubit", wires=qubits)
 
@@ -103,29 +104,34 @@ def test_batch_execution_of_gradient_torch(device, shots, mocker):
         qml.templates.StronglyEntanglingLayers(weights, wires=range(qubits))
         return qml.expval(qml.PauliZ(0))
 
-    qnode_aws = qml.QNode(func, dev_aws, interface="torch", diff_method="parameter-shift")
+    qnode_braket = qml.QNode(func, dev_braket, interface="torch", diff_method="parameter-shift")
     qnode_default = qml.QNode(func, dev_default, interface="torch", diff_method="parameter-shift")
 
     shape = qml.templates.StronglyEntanglingLayers.shape(layers, qubits)
     weights = np.random.random(shape)
-    weights_aws = torch.tensor(weights, requires_grad=True)
+    weights_braket = torch.tensor(weights, requires_grad=True)
     weights_default = torch.tensor(weights, requires_grad=True)
 
-    spy1 = mocker.spy(BraketAwsQubitDevice, "execute")
-    spy2 = mocker.spy(BraketAwsQubitDevice, "batch_execute")
-    spy3 = mocker.spy(AwsDevice, "run_batch")
+    if isinstance(dev_braket, BraketAwsQubitDevice):
+        spy1 = mocker.spy(BraketAwsQubitDevice, "execute")
+        spy2 = mocker.spy(BraketAwsQubitDevice, "batch_execute")
+        spy3 = mocker.spy(AwsDevice, "run_batch")
+    elif isinstance(dev_braket, BraketLocalQubitDevice):
+        spy1 = mocker.spy(BraketLocalQubitDevice, "execute")
+        spy2 = mocker.spy(BraketLocalQubitDevice, "batch_execute")
+        spy3 = mocker.spy(LocalSimulator, "run_batch")
 
-    out_aws = qnode_aws(weights_aws)
+    out_braket = qnode_braket(weights_braket)
     out_default = qnode_default(weights_default)
 
-    out_aws.backward()
+    out_braket.backward()
     out_default.backward()
 
-    res_aws = weights_aws.grad
+    res_braket = weights_braket.grad
     res_default = weights_default.grad
 
     if qml.version() >= "0.20.0":
-        assert np.allclose(res_aws, res_default)
+        assert np.allclose(res_braket, res_default)
         spy1.assert_not_called()
         assert len(spy2.call_args_list) == 2
         assert len(spy3.call_args_list) == 2
@@ -136,7 +142,7 @@ def test_batch_execution_of_gradient_torch(device, shots, mocker):
             len(spy2.call_args_list[1][0][1]) == expected_circuits
         )  # Then called for backward pass
     else:
-        assert np.allclose(res_aws, res_default)
+        assert np.allclose(res_braket, res_default)
         spy1.assert_called_once()  # For a forward pass
         spy2.assert_called_once()
         spy3.assert_called_once()
@@ -154,12 +160,9 @@ def test_batch_execution_of_gradient_tf(device, shots, mocker):
     qubits = 2
     layers = 2
 
-    dev_aws = device(qubits)
+    dev_braket = device(qubits)
 
-    if isinstance(dev_aws, BraketLocalQubitDevice):
-        pytest.skip("Parallelized batch execution is only supported on the remote AWS device")
-
-    dev_aws._parallel = True
+    dev_braket._parallel = True
 
     dev_default = qml.device("default.qubit", wires=qubits)
 
@@ -167,22 +170,27 @@ def test_batch_execution_of_gradient_tf(device, shots, mocker):
         qml.templates.StronglyEntanglingLayers(weights, wires=range(qubits))
         return qml.expval(qml.PauliZ(0))
 
-    qnode_aws = qml.QNode(func, dev_aws, interface="tf", diff_method="parameter-shift")
+    qnode_braket = qml.QNode(func, dev_braket, interface="tf", diff_method="parameter-shift")
     qnode_default = qml.QNode(func, dev_default, interface="tf", diff_method="parameter-shift")
 
     shape = qml.templates.StronglyEntanglingLayers.shape(layers, qubits)
     weights = np.random.random(shape)
-    weights_aws = tf.Variable(weights)
+    weights_braket = tf.Variable(weights)
     weights_default = tf.Variable(weights)
 
-    spy1 = mocker.spy(BraketAwsQubitDevice, "execute")
-    spy2 = mocker.spy(BraketAwsQubitDevice, "batch_execute")
-    spy3 = mocker.spy(AwsDevice, "run_batch")
+    if isinstance(dev_braket, BraketAwsQubitDevice):
+        spy1 = mocker.spy(BraketAwsQubitDevice, "execute")
+        spy2 = mocker.spy(BraketAwsQubitDevice, "batch_execute")
+        spy3 = mocker.spy(AwsDevice, "run_batch")
+    elif isinstance(dev_braket, BraketLocalQubitDevice):
+        spy1 = mocker.spy(BraketLocalQubitDevice, "execute")
+        spy2 = mocker.spy(BraketLocalQubitDevice, "batch_execute")
+        spy3 = mocker.spy(LocalSimulator, "run_batch")
 
     with tf.GradientTape() as tape:
-        out_aws = qnode_aws(weights_aws)
+        out_braket = qnode_braket(weights_braket)
 
-    res_aws = tape.gradient(out_aws, weights_aws)
+    res_braket = tape.gradient(out_braket, weights_braket)
 
     with tf.GradientTape() as tape:
         out_default = qnode_default(weights_default)
@@ -190,7 +198,7 @@ def test_batch_execution_of_gradient_tf(device, shots, mocker):
     res_default = tape.gradient(out_default, weights_default)
 
     if qml.version() >= "0.20.0":
-        assert np.allclose(res_aws, res_default)
+        assert np.allclose(res_braket, res_default)
         spy1.assert_not_called()
         assert len(spy2.call_args_list) == 2
         assert len(spy3.call_args_list) == 2
@@ -201,7 +209,7 @@ def test_batch_execution_of_gradient_tf(device, shots, mocker):
             len(spy2.call_args_list[1][0][1]) == expected_circuits
         )  # Then called for backward pass
     else:
-        assert np.allclose(res_aws, res_default)
+        assert np.allclose(res_braket, res_default)
         spy1.assert_called_once()  # For a forward pass
         spy2.assert_called_once()
         spy3.assert_called_once()
