@@ -17,7 +17,10 @@ import pennylane as qml
 import pytest
 from pennylane import numpy as np
 
-from braket.pennylane_plugin.braket_device import MIN_SIMULATOR_BILLED_MS, BraketAwsQubitDevice
+from braket.pennylane_plugin.braket_device import (
+    MIN_SIMULATOR_BILLED_MS,
+    BraketAwsQubitDevice,
+)
 
 
 @pytest.mark.parametrize("shots", [100])
